% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{calculate_celltype_prob}
\alias{calculate_celltype_prob}
\title{calculate_celltype_prob}
\usage{
calculate_celltype_prob(clt_marker_list, marker_database_list, type = "jacc")
}
\arguments{
\item{clt_marker_list}{A list of markers for all cluster.}

\item{marker_database_list}{A list of markers of all reference cell types.}

\item{type}{A parameter to select the method to measure cluster and cell type
similarity
\itemize{
\item jacc - Jaccard index.
\item ac - Accuracy.
\item f1 - F1 score.
}.}
}
\value{
A confusion matrix between clusters and cell types. Each cell represents
a probabilty of a cluster belongs to a cell type.
}
\description{
Calculate clusters and cell types similarity based on the markers.
}
