% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{curate_markers}
\alias{curate_markers}
\title{curate_markers}
\usage{
curate_markers(
  whole_list,
  gene_names,
  wilcox_threshold = 0.001,
  logfc_threshold = 1.5
)
}
\arguments{
\item{whole_list}{A list of markers for all clusters.}

\item{gene_names}{All the gene names of the expression matrix.}

\item{wilcox_threshold}{A threshold for p-value \code{wilcox_threshold = 0.001} by default.}

\item{logfc_threshold}{A threshold for fold-change \code{logfc_threshold = 1.5} by default.}
}
\value{
A list of markers that are strong expressed for discovered clusters.
}
\description{
Filter genes that have low p-value and fold-change.
}
