% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{find_markers}
\alias{find_markers}
\title{find_markers}
\usage{
find_markers(input_data_matrix, cluster_labels, identity = 1, threads = 8)
}
\arguments{
\item{input_data_matrix}{An expression matrix in which rows are genes and
columns are cells.}

\item{cluster_labels}{A vector of cluster labels obtained from clustering methods.}

\item{identity}{A parameter to select specific cluster \code{identity = 1}
by default.}

\item{threads}{A parameter to control number of cores used for analysis
\code{threads = 1} by default.}
}
\value{
A list that contains p-value and fold-change ratio for all genes of each cluster.
}
\description{
Perform cluster-wise Wilcox test and fold-change for each gene.
}
