% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{find_specific_marker}
\alias{find_specific_marker}
\title{find_specific_marker}
\usage{
find_specific_marker(gene_name, f_list, type = "jacc")
}
\arguments{
\item{gene_name}{A list of markers belong to the cluster.}

\item{f_list}{A list of markers belongs to a reference cell type.}

\item{type}{A parameter to select the method to measure cluster and cell type
similarity
\itemize{
\item jacc - Jaccard index.
\item ac - Accuracy.
\item f1 - F1 score.
}.}
}
\value{
A vector of probabilties of a cluster belongs to cell types.
}
\description{
Calculate cluster and cell type similarity based on the markers.
}
