% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_cluster_markers}
\alias{get_cluster_markers}
\title{get_cluster_markers}
\usage{
get_cluster_markers(input_data_matrix, labels_vector, threads = 1)
}
\arguments{
\item{input_data_matrix}{An expression matrix in which rows are genes and
columns are cells.}

\item{labels_vector}{A vector of cluster labels ontained from clustering methods.}

\item{threads}{A parameter to control number of cores used for analysis
\code{threads = 1} by default.}
}
\value{
A list that contains markers for each cluster.
}
\description{
Find markers for each cluster
}
