% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCAN.R
\docType{data}
\name{SCE}
\alias{SCE}
\title{SCE}
\format{
An object of class \code{list} of length 2.
}
\usage{
SCE
}
\description{
SCE dataset includes scRNA-seq data and cell type information.
}
\keyword{datasets}
