% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{adjustedRandIndex}
\alias{adjustedRandIndex}
\title{adjustedRandIndex}
\usage{
adjustedRandIndex(x, y)
}
\arguments{
\item{x}{A vector that contain predicted cluster assignment.}

\item{y}{A vector that contain true cluster assignment.}
}
\description{
The function to calculate adjusted Rand index value with the
inputs of true clusters and predicted clusters
}
