% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCAN.R
\name{scCAN}
\alias{scCAN}
\title{scCAN}
\usage{
scCAN(
  data,
  sparse = F,
  n.neighbors = 30,
  alpha = 0.5,
  n.iters = 10,
  ncores = 10L,
  r.seed = 1
)
}
\arguments{
\item{data}{Gene expression matrix, with rows represent samples and columns represent genes.}

\item{sparse}{Boolen variable indicating whether data is a sparse matrix. The input must be a non negative sparse matrix.}

\item{n.neighbors}{Number of neighboring cells that are used to caculate the edge's weight. The number of neighbors are set \code{n.neighbors = 30} by default.}

\item{alpha}{A hyper-parameter that is used to calculate the network kernel. The value is set to \code{alpha = 0.5} by default.}

\item{n.iters}{A hyper-parameter to set the number of network fusion iterations. It is set to \code{n.iters = 10} by default.}

\item{ncores}{Number of processor cores to use.}

\item{r.seed}{A parameter to set a seed for reproducibility. This values is set to \code{r.seed = 1} by default.}
}
\value{
List with the following keys:
\itemize{
\item cluster - A numeric vector containing cluster assignment for each sample. If \code{do.clus = False}, this values is always \code{NULL}.
\item latent - A matrix representing compressed data from the input data, with rows represent samples and columns represent latent variables.
}
}
\description{
This is the main function to perform sc-RNA seq data clustering clustering. scCAN is fully unsupervised scRNA-seq
clustering framework that uses deep neural network and network fusion-based clustering algorithm.
First, scCAN applies a non-negative autoencoder to filter scRNA-seq data.
Second, the filtered data is passed to stacked Bayesian autoencoder to get multiple low-dimensional representations of input data.
Subsequently, scCAN converts these compressed data into networks and unify those networks to a single graph.
Then, scCAN uses a spectral clustering algorithm to obtain final clusters assignment.
}
